Add an efficient strategy sampler to help with debugging
It is useful, when developing strategies, to have a way of measuring the performance of strategies asymptotically as a function of the number of examples you want to generate. The function debug_strategy is designed to sample n examples from a strategy, without incurring the added overhead that a naive approach would take i.e. using a for loop and calling strategy.example() n times.