Separated debug_strategy to distinguish behavior on draw() vs example()
The behavior on draw() should not scale dramatically, as you pay overhead for larger sample sizes, which we start to feel at around 10,000. On the other hand, the behavior on example() should scale, as it doesn't pay the same overhead as draw() (e.g. for shrinking), and it should easily hit 100,000 samples.